AP_Motors: convert to 32 bit motor mask

+++ libraries/AP_Motors/AP_MotorsCoax.h
@@ -47,7 +47,7 @@ public:
 
     // get_motor_mask - returns a bitmask of which outputs are being used for motors or servos (1 means being used)
     //  this can be used to ensure other pwm outputs (i.e. for servos) do not conflict
-    uint16_t            get_motor_mask() override;
+    uint32_t            get_motor_mask() override;
 
     const char* get_frame_string() const override { return "COAX"; }
